chore(api): move generate-username controller  into prescription context
