+++ api/src/certification/session-management/domain/services/session-publication-service.js
@@ -1,8 +1,5 @@
 /**
  * @typedef {import('../../../../../src/certification/session-management/domain/usecases/index.js').CertificationRepository} CertificationRepository
- * @typedef {import('../../../../../lib/domain/usecases/index.js').FinalizedSessionRepository} FinalizedSessionRepository
- * @typedef {import('../../../../../lib/domain/usecases/index.js').SessionRepository} SessionRepository
- * @typedef {import('../../../../../lib/domain/usecases/index.js').SharedSessionRepository} SharedSessionRepository
  * @typedef {import('../../../../../src/certification/session-management/domain/usecases/index.js').MailService} MailService
  */
 import lodash from 'lodash';
