+++ api/src/prescription/organization-learner/application/sco-organization-learner-controller.js
@@ -128,12 +128,35 @@ const generateUsernameWithTemporaryPassword = async function (
   return h.response(dependencies.scoOrganizationLearnerSerializer.serializeCredentialsForDependent(result)).code(200);
 };
 
+const generateUsername = async function (request, h, dependencies = { scoOrganizationLearnerSerializer }) {
+  const payload = request.payload.data.attributes;
+  const { 'organization-id': organizationId } = payload;
+
+  const studentInformation = {
+    firstName: payload['first-name'],
+    lastName: payload['last-name'],
+    birthdate: payload['birthdate'],
+  };
+
+  const username = await usecases.generateUsername({ organizationId, studentInformation });
+
+  const scoOrganizationLearner = {
+    ...studentInformation,
+    username,
+  };
+
+  return h
+    .response(dependencies.scoOrganizationLearnerSerializer.serializeWithUsernameGeneration(scoOrganizationLearner))
+    .code(200);
+};
+
 const scoOrganizationLearnerController = {
   createUserAndReconcileToOrganizationLearnerFromExternalUser,
   createAndReconcileUserToOrganizationLearner,
   updatePassword,
   batchGenerateOrganizationLearnersUsernameWithTemporaryPassword,
   generateUsernameWithTemporaryPassword,
+  generateUsername,
 };
 
 export { scoOrganizationLearnerController };
